Connect to a remote server

Connect to a remote server

get_connection(host, user = tolower(paste0(Sys.getenv("USER"),
  Sys.getenv("USERNAME"))), password, port = 22)

Arguments

host
the name of the remote system

user
a string for the username on the remote system. The default is to use the same name as the current user.

password
the password for the user on the remote system

port
this is usually 22 for ssh connections

Value

an object of class Rsch_connection

Examples

# NOT RUN {
connection <- get_connection(host = "cuckoo.lbnl.gov", user = "cliff", password = "TopSecret")
# }